refactor(gui): extract DaemonTailPane into a shared observe component
DaemonObserveView owned the only observe.tail pane in the app, unexported, so a
second mount point had no way to reuse its polling loop, cursor handling, pause
semantics and row rendering.

Pure move: the pane, its row view, the tail hook and their constants relocate to
components/observe/DaemonTailPane.tsx and the view imports them. No behaviour
change -- daemon-observe.vitest.ts and entity-id-links.vitest.ts pass unmodified.
